support raw headers format
according to the node docs:

headers may be an Array where the keys and values are in the same list.
It is not a list of tuples. So, the even-numbered offsets are key
values, and the odd-numbered offsets are the associated values. The
array is in the same format as request.rawHeaders.

we still support Array of tupples for backwards compatibility